requirements.txt
# Core web scraping dependencies
requests>=2.32.0
beautifulsoup4>=4.13.0
lxml>=6.0.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
python-dateutil>=2.9.0
typing-extensions>=4.15.0
unidecode>=1.4.0
urllib3>=2.5.0
certifi>=2025.8.3
charset-normalizer>=3.4.3
idna>=3.10
pytest>=7.0.0
pytest-cov>=4.0.0
# optional
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0